Created JUnit test
I think game logic should be a static method because as of right now I
don't know how to access its methods for testing. The tests for win/loss
will be written once I get a better idea of how those states will be
implemented. The testing code is a bit long but I have included a .jpg
that illustrates the node board I used for testing.

I also changed the ticket values for Detective and MrX because bus and
taxi were swapped. Also fixed the filename for PlayerID.